TECOPS-5891 Register Databricks schema
